Redis search-queue and search-worker refactoring
