+++ types/search-queue.ts
@@ -1,0 +1,8 @@
+import { FilterValue } from './search-filters';
+
+export interface SearchQueueInterface {
+  query: string;
+  k: number;
+  user_id: string;
+  filters: FilterValue[];
+}
